+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-- Added support for `animationSingleArmAnimation` and `animationDontShowArmor`.
-- Legacy4J: Right Stick Up and Right Shift now cycle skin poses in reverse.
-- Fixed a potential `@Redirect` conflict in `MixinCapeFeatureRenderer`.
+- Player head icons on screens such as the player list are now replaced by the selected skin.
+- Updated Elytra rendering logic for skins with capes.
 
 **Full Changelog**: https://github.com/BrandonItaly/Bedrock-Skins/commits/master